ENH: Add -o|--output parameter on Linux
That parameter redirects all script output into a FILESPEC or &N file descriptor. Piping the entire output of runVimTests is problematic because the started Vim instances expect to write the UI to stdout; without that, the screen does not update / is messed up, and you cannot do debugging in there.